Formally drops support for Python 2.6 and Python 3.3.
These have been deprecated and unsupported for a while, but
haven't caused problems. The Python 3.3 test is now causing
Travis CI suite to fail and, while that isn't a great reason on
its own to drop support, it's the final nail in the coffin. This
had to happen at some point, and this is that point. On the plus
side, we can now move forward with newer versions of pytest.
Yay!